Transform app to Origin Client with bottom navigation and fragments
