+++ app/build.gradle
@@ -4,11 +4,11 @@ plugins {
 }
 
 android {
-    namespace 'io.bambosan.mbloader'
+    namespace 'com.origin.launcher'
     compileSdk 35
     buildToolsVersion = "35.0.0"
     defaultConfig {
-        applicationId "io.bambosan.mbloader"
+        applicationId "com.origin.launcher"
         minSdk 26
         targetSdk 35
         versionCode 11
