+++ app/src/main/res/layout/fragment_home.xml
@@ -1,0 +1,86 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout
+	xmlns:android="http://schemas.android.com/apk/res/android"
+	xmlns:app="http://schemas.android.com/apk/res-auto"
+	android:id="@+id/linearlayout"
+	android:layout_width="match_parent"
+	android:layout_height="match_parent"
+	android:orientation="vertical"
+	android:padding="8dp">
+
+	<TextView
+		android:id="@+id/textview"
+		android:layout_width="match_parent"
+		android:layout_height="wrap_content"
+		android:text="@string/welcome"
+		android:textSize="19sp"
+		android:gravity="center_horizontal" />
+
+	<LinearLayout
+		android:id="@+id/tablelayout"
+		android:layout_width="match_parent"
+		android:layout_height="wrap_content"		
+		android:orientation="vertical"
+		android:padding="8dp"
+		android:background="#FF000000">
+
+		<TextView
+			android:id="@+id/textview1"
+			android:layout_width="wrap_content"
+			android:layout_height="wrap_content"
+			android:text="@string/mbl2_desc" />
+
+		<Button
+			android:id="@+id/mbl2_load"
+			android:layout_width="wrap_content"
+			android:layout_height="wrap_content"
+			android:text="Load with mbl2" />
+
+	</LinearLayout>
+
+	<LinearLayout
+		android:id="@+id/linearlayout1"
+		android:layout_width="match_parent"
+		android:layout_height="wrap_content"
+		android:orientation="horizontal"
+		android:padding="8dp">
+
+		<TextView
+			android:id="@+id/textview3"
+			android:layout_width="wrap_content"
+			android:layout_height="wrap_content"
+			android:text="@string/mc_pkgname" />
+
+		<EditText
+			android:id="@+id/mc_pkgname"
+			android:layout_width="wrap_content"
+			android:layout_height="wrap_content"
+			android:hint="Package name of Minecraft"
+			android:text="com.mojang.minecraftpe"
+			android:inputType="text" />
+
+	</LinearLayout>
+
+	<TableLayout
+		android:id="@+id/tablelayout2"
+		android:layout_width="match_parent"
+		android:layout_height="wrap_content"
+		android:padding="8dp">
+
+		<TextView
+			android:id="@+id/textview4"
+			android:layout_width="match_parent"
+			android:layout_height="wrap_content"
+			android:text="Launcher logs here:"
+			android:gravity="center_horizontal"
+			android:textSize="19sp" />
+
+		<TextView
+			android:id="@+id/listener"
+			android:layout_width="wrap_content"
+			android:layout_height="wrap_content"
+			android:text="   " />
+
+	</TableLayout>
+
+</LinearLayout>
